Fixes TXT and CAA records (#7)
TL;DR
-----

Updates the data in the TXT and CAA records to prevent spurious
mismatches

Details
-------

The value of the spam verification and CAA records that applied
correctly but replied on GCP making an update to put the right
quotes into the record. This led to spurious "changes" on plan/apply.
Updates the value to only show real changes.